+++ addon/components/polaris-page/header.js
@@ -26,6 +26,16 @@ export default Component.extend({
   title: null,
 
   /**
+   * Important and non-interactive status information shown immediately after the title
+   *
+   * @property titleMetadata
+   * @public
+   * @type {String|Component}
+   * @default null
+   */
+  titleMetadata: null,
+
+  /**
    * Visually hide the title
    *
    * @property titleHidden
@@ -115,6 +125,18 @@ export default Component.extend({
   hasActions: or('primaryAction', 'secondaryActions').readOnly(),
   hasSecondaryActions: gt('secondaryActions.length', 0).readOnly(),
   hasActionGroups: gt('actionGroups.length', 0).readOnly(),
+
+  shouldRenderPrimaryActionAsPrimary: computed(
+    'primaryAction.primary',
+    function() {
+      let primaryAction = this.get('primaryAction');
+
+      return (
+        primaryAction &&
+        (primaryAction.primary === undefined ? true : primaryAction.primary)
+      );
+    }
+  ).readOnly(),
 
   hasRollup: computed(
     'secondaryActions.length',