+++ addon/components/polaris-text-style.js
@@ -1,16 +1,38 @@
 import Component from '@ember/component';
 import { computed } from '@ember/object';
-import { isEmpty } from '@ember/utils';
 import { classify } from '@ember/string';
 import layout from '../templates/components/polaris-text-style';
+
+const variationValue = {
+  positive: 'positive',
+  negative: 'negative',
+  strong: 'strong',
+  subdued: 'subdued',
+  code: 'code',
+};
+
+const variationElement = (variation) => {
+  switch (variation) {
+    case variationValue.code:
+      return 'code';
+    case variationValue.strong:
+      return 'b';
+    case variationValue.positive:
+    case variationValue.negative:
+    case variationValue.subdued:
+    default:
+      return 'span';
+  }
+};
 
 /**
  * Polaris text style component.
  * See https://polaris.shopify.com/components/titles-and-text/text-style
+ *
+ * @component polaris-text-style
  */
 export default Component.extend({
-  tagName: 'span',
-  classNameBindings: ['variationClass'],
+  tagName: '',
 
   layout,
 
@@ -40,12 +62,21 @@ export default Component.extend({
    */
   text: null,
 
+  dataTestTextStyle: true,
+
+  /**
+   * @private
+   */
+  elementTagName: computed('variation', function() {
+    return variationElement(this.get('variation'));
+  }).readOnly(),
+
   /**
    * @private
    */
   variationClass: computed('variation', function() {
     const variation = this.get('variation');
-    if (isEmpty(variation)) {
+    if (!Object.keys(variationValue).includes(variation)) {
       return null;
     }